requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
plotly>=5.18.0
rich>=13.7.0
